+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Fix issue with `useViewModelCache` that could cause a render after unmount
 * (breaking) Removed `RequestError` from `Endpoint`. Instead `Endpoint` will now throw the error raised by `fetch` directly - either `TypeError` or `AbortError`.
 * Fix bugs with `RelatedViewModelField` and `ManyRelatedViewModelField` where things could break if value was `null` or `[]`
+* Fix bug with `ViewModelCache` when passing a record with nested related fields to `get` or `getList` that meant it only traversed 1 level deep which meant the returned record could be different to the passed record.
 
 ## [0.0.11] - 2021-03-01